Fix clipboard-related event handlers
This fixes a bug where the clipboard was getting overwritten by server-side copies, even when `clipboard_direction=to-server`. This prevents server-side copies when `clipboard_direction=to-server`.

One must set `clipboard_direction` in `default_settings.txt` to use this fix.

Note: We don't need to send the clipboard on every keyboard event. Capturing the contents on focus should be sufficient, if the intent is to copy data from the client browser to the server.