fix broken legacy clipboard
The DataTransfer object hold by clipboardData has an empty list of
files. Due do an empty list being a valid list, the code path for file
transfers is used, which breaks copy'n'paste of text.

The issue can be observed with Firefox (asyncClipboard support still
disabled by default) as well as Chrome in case a HTTP connection is
used.