Bugfix: Prefix the sessionStorage data with pathname
In case of multiple xpra sessions behind a single domain (i.e. reverse proxy), the sessions will overwrite their settings inside the sessionStorage
when the user navigates between the sessions within a browser tab.
This change will store the sessionStorage data prefixed with the pathname of the xpra session.
This fixes #278